CSV output is now done using an RFC4180 library and there is an option when saving as CSV to duplicate each term once per document so you can see how many times the term appeared in each document.

+++ src/main/java/gate/termraider/gui/SliderPanel.java
@@ -44,7 +44,7 @@ public class SliderPanel extends JPanel {
     if (minScore instanceof Integer) {
       imin = minScore.intValue();
       imax = this.scoredbank.getMaxScore().intValue();
-      middle = (int) ( (imin + imax) / 2);
+      middle = (imin + imax) / 2;
     }
 
     // This is for everything else
